Refactor `FfiResult`

+++ crates/capi/src/refcount.rs
@@ -18,7 +18,7 @@ pub extern "C" fn _Py_IncRef(op: *mut PyObject) {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn Py_REFCNT(op: *mut PyObject) -> isize {
-    with_vm(|_vm| unsafe { &*op }.strong_count() as isize)
+    with_vm(|_vm| unsafe { &*op }.strong_count())
 }
 
 #[cfg(test)]
